timed task is now on node server and is on a worker thread. * Migrated all code to node server * Changed config.json * Sending of guild PVP is using webhooks (up to Odious if he wants to keep)
